+++ src/org/netbeans/modules/php/blade/editor/lexer/BladeTokenId.java
@@ -76,28 +76,69 @@ public enum BladeTokenId implements TokenId {
                 LanguagePath languagePath, InputAttributes inputAttributes) {
             boolean joinHtml = true;
             switch (token.id()) {
-
+                case PHP_BLADE_INLINE_CODE:
                 case PHP_BLADE_EXPRESSION: {
                     Language<? extends TokenId> phpLanguage = PHPTokenId.languageInPHP();
+                    if (phpLanguage == null || token.text() == null){
+                        return null;
+                    }
+
+                    //php brace matcher freeze issue patch
                     String tokenText = token.text().toString();
                     int startOffset = 1;
                     int endOffset = 1;
 
-                    //temprary corner case
-                    if (tokenText.startsWith("((") && tokenText.endsWith("))")){
+                    if (tokenText.startsWith("((") && tokenText.endsWith("))")){ //NOI18N
                         startOffset = 2;
                         endOffset = 2;
+                    } else if (tokenText.startsWith("([") && tokenText.endsWith("])")){ //NOI18N
+                        startOffset = 2;
+                        endOffset = 2;
+                    } else if (tokenText.startsWith("([")){ //NOI18N
+                        startOffset = 2;
                     }
-                    return phpLanguage != null ? LanguageEmbedding.create(phpLanguage, startOffset, endOffset, false) : null;
+                    return LanguageEmbedding.create(phpLanguage, startOffset, endOffset, false);
                 }
                 case PHP_BLADE_ECHO_EXPR:
-                case PHP_BLADE_INLINE_CODE:
-                    /**
-                     * troubleshooting php embedding freeze (?:, ::) force a
-                     * return null;
-                     */
                     Language<? extends TokenId> phpLanguage = PHPTokenId.languageInPHP();
-                    return phpLanguage != null ? LanguageEmbedding.create(phpLanguage, 0, 0, false) : null;
+                    if (phpLanguage == null || token.text() == null){
+                        return null;
+                    }
+                    
+                    String tokenText = token.text().toString();
+                    int startOffset = 0;
+                    int endOffset = 0;
+                    
+                    if (!tokenText.startsWith("(") && !tokenText.startsWith("[")){
+                        return LanguageEmbedding.create(phpLanguage, startOffset, endOffset, false);
+                    }
+
+                    //php brace matcher freeze issue patch
+                    if (tokenText.startsWith("((") && tokenText.endsWith("))")){ //NOI18N
+                        startOffset = 2;
+                        endOffset = 2;
+                    } else if (tokenText.startsWith("[[") && tokenText.endsWith("]]")){ //NOI18N
+                        startOffset = 2;
+                        endOffset = 2;
+                    } else if (tokenText.startsWith("([") && tokenText.endsWith("])")){ //NOI18N
+                        startOffset = 2;
+                        endOffset = 2;
+                    } else if (tokenText.startsWith("[(") && tokenText.endsWith(")]")){ //NOI18N
+                        startOffset = 2;
+                        endOffset = 2;
+                    } else if (tokenText.startsWith("([") || tokenText.startsWith("[(")){ //NOI18N
+                        startOffset = 2;
+                    } else if (tokenText.startsWith("(") && tokenText.endsWith(")")){ //NOI18N
+                        startOffset = 1;
+                        endOffset = 1;
+                    } else if (tokenText.startsWith("[") && tokenText.endsWith("]")){ //NOI18N
+                        startOffset = 1;
+                        endOffset = 1;
+                    }  else if (tokenText.startsWith("(") || tokenText.startsWith("[")){ //NOI18N
+                        startOffset = 1;
+                    }
+                    
+                    return LanguageEmbedding.create(phpLanguage, startOffset, endOffset, false);
                 case PHP_INLINE:
                     Language<? extends TokenId> phpLanguageCode = PHPTokenId.language();
                     return phpLanguageCode != null ? LanguageEmbedding.create(phpLanguageCode, 0, 0, false) : null;